--- project/src/components/not-logged/not-logged.tsx
@@ -1,20 +1,0 @@
-import {Link} from 'react-router-dom';
-import {AppRoute} from '../../const';
-
-function NotLogged(): JSX.Element {
-  return (
-    <nav className="header__nav">
-      <ul className="header__nav-list">
-        <li className="header__nav-item user">
-          <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Login}>
-            <div className="header__avatar-wrapper user__avatar-wrapper">
-            </div>
-            <span className="header__login">Sign in</span>
-          </Link>
-        </li>
-      </ul>
-    </nav>
-  );
-}
-
-export default NotLogged;
