+++ project/src/pages/property-page/property-page.tsx
@@ -1,5 +1,5 @@
 import Header from '../../components/header/header';
-import NotLogged from '../../components/not-logged/not-logged';
+import HeaderNav from '../../components/header-nav/header-nav';
 import {Offers} from '../../types/offer';
 import { Reviews } from '../../types/review';
 import {useParams} from 'react-router-dom';
@@ -31,7 +31,7 @@ function PropertyPage(props: PropertyPageProps): JSX.Element {
   return (
     <div className="page">
       <Header>
-        <NotLogged/>
+        <HeaderNav/>
       </Header>
       <main className="page__main page__main--property">
         <section className="property">
