+++ project/src/store/action.ts
@@ -11,5 +11,5 @@ export const loadOffers = createAction<Offers>('loadOffers');
 export const setDataLoadingStatus = createAction<boolean>('setDataLoadingStatus');
 export const requireAuthorization = createAction<AuthorizationStatus>('requireAuthorisation');
 export const redirectToRoute = createAction<AppRoute>('redirectToRoute');
-export const loadUserData = createAction<UserData>('loadUserData');
+export const loadUserData = createAction<UserData | undefined>('loadUserData');
