+++ project/src/store/middlewares/redirect.ts
@@ -8,7 +8,7 @@ export const redirect: Middleware<unknown, Reducer> =
   (_store) =>
     (next) =>
       (action) => {
-        if (action.type === 'game/redirectToRoute') {
+        if (action.type === 'redirectToRoute') {
           browserHistory.push(action.payload);
         }
